requirements.txt
langgraph>=0.0.60
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.1
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
